requirements.txt
streamlit
transformers
torch
pillow
ollama